# spring/docs/openapi.yaml
openapi: "3.0.1"
info:
  title: "Spring部分的Swagger文档"
  description: "这是项目的Spring部分的Swagger文档"
  version: "1.0.0"
...
paths:
...
  /articles:
    put:
      tags:
      - "文章模块"
      summary: "更新文章"
      description: "根据DTO更新文章信息"
      operationId: "updateArticle"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdateDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
    post:
      tags:
      - "文章模块"
      summary: "创建文章"
      description: "通过请求体创建一篇新文章"
      operationId: "createArticle"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleCreateDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultLong"
  /articles/view/{id}:
    put:
      tags:
      - "文章模块"
      summary: "增加文章阅读量"
      description: "增加文章阅读量"
      operationId: "addViewArticle"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
  /articles/publish/{id}:
    put:
      tags:
      - "文章模块"
      summary: "发布文章"
      description: "将文章状态修改为发布"
      operationId: "publishArticle"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
...
  /articles/views/batch:
    post:
      tags:
      - "文章模块"
      summary: "批量查询文章阅读量（内部）"
      description: "根据ID列表批量查询文章阅读量，供内部服务远程调用"
      operationId: "getArticleViewsByIds"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIdsDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListIdCountVO"
  /articles/batch:
    post:
      tags:
      - "文章模块"
      summary: "批量查询文章（内部）"
      description: "根据ID列表批量查询文章，供内部服务远程调用"
      operationId: "getArticleByIds"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIdsDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListArticleWithCategoryVO"
...
  /comments/article/{id}:
    get:
      tags:
      - "评论模块"
      summary: "根据文章id分页获取评论"
      description: "根据文章id分页获取评论"
      operationId: "listCommentsByArticleId"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "sort_way"
        in: "query"
        required: false
        schema:
          type: "string"
          default: "create_time"
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOCommentsVO"
...
  /articles/{id}:
    get:
      tags:
      - "文章模块"
      summary: "获取文章详情"
      description: "根据ID获取文章详情"
      operationId: "getArticleById"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultArticleWithCategoryVO"
    delete:
      tags:
      - "文章模块"
      summary: "删除文章"
      description: "根据ID删除文章"
      operationId: "deleteArticle"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
  /articles/user/{id}:
    get:
      tags:
      - "文章模块"
      summary: "获取用户文章"
      description: "返回用户文章，可指定是否只查询已发布的文章"
      operationId: "getArticlesByUserId"
      parameters:
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "published"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 0
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOArticleWithCategoryVO"
  /articles/unpublished/list:
    get:
      tags:
      - "文章模块"
      summary: "获取所有未发布文章"
      description: "返回所有未发布的文章，支持分页"
      operationId: "getUnpublishedArticles"
      parameters:
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOArticleWithCategoryVO"
...
  /articles/neo4j-sync:
    get:
      tags:
      - "文章模块"
      summary: "获取文章表数据用于Neo4j同步（内部）"
      description: "获取文章表数据，支持增量同步，供FastAPI同步Neo4j使用"
      operationId: "getNeo4jSyncArticles"
      parameters:
      - name: "updated_after"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListArticleSyncVO"
  /articles/list:
    get:
      tags:
      - "文章模块"
      summary: "获取文章列表"
      description: "返回所有已发布的文章"
      operationId: "getPublishedArticles"
      parameters:
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOArticle"
  /articles/by-title:
    get:
      tags:
      - "文章模块"
      summary: "根据标题模糊搜索文章（内部）"
      description: "根据标题模糊搜索，供内部服务远程调用"
      operationId: "getArticlesByTitle"
      parameters:
      - name: "title"
        in: "query"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListArticleWithCategoryVO"
...
  /articles/batch/{ids}:
    delete:
      tags:
      - "文章模块"
      summary: "批量删除文章"
      description: "根据ID数组批量删除文章，多个ID用英文逗号分隔"
      operationId: "deleteArticles"
      parameters:
      - name: "ids"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
components:
  schemas:
...
    ResultVoid:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "null"
          description: "返回数据"
          nullable: true
      description: "统一响应"
...
    ArticleUpdateDTO:
      required:
      - "content"
      - "id"
      - "status"
      - "sub_category_id"
      - "title"
      - "username"
      type: "object"
      properties:
        title:
          maxLength: 200
          minLength: 1
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章内容"
        username:
          maxLength: 50
          minLength: 1
          type: "string"
          description: "作者用户名"
        tags:
          maxLength: 200
          minLength: 0
          type: "string"
          description: "标签，多个用英文逗号分隔"
        status:
          maximum: 1
          minimum: 0
          type: "integer"
          description: "状态，0草稿 1已发布"
          format: "int32"
        sub_category_id:
          minimum: 1
          type: "integer"
          description: "子分类ID"
          format: "int32"
        id:
          minimum: 1
          type: "integer"
          description: "文章ID"
          format: "int64"
      description: "更新文章请求"
...
    ResultLong:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "integer"
          description: "返回数据"
          format: "int64"
      description: "统一响应"
...
    BatchIdsDTO:
      required:
      - "ids"
      type: "object"
      properties:
        ids:
          type: "array"
          description: "ID列表"
          items:
            type: "integer"
            description: "ID列表"
            format: "int64"
      description: "批量ID请求"
...
    IdCountVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "ID"
          format: "int64"
        count:
          type: "integer"
          description: "数量"
          format: "int64"
      description: "ID 与数量"
...
    ArticleCreateDTO:
      required:
      - "content"
      - "status"
      - "sub_category_id"
      - "title"
      - "username"
      type: "object"
      properties:
        title:
          maxLength: 200
          minLength: 1
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章内容"
        username:
          maxLength: 50
          minLength: 1
          type: "string"
          description: "作者用户名"
        tags:
          maxLength: 200
          minLength: 0
          type: "string"
          description: "标签，多个用英文逗号分隔"
        status:
          maximum: 1
          minimum: 0
          type: "integer"
          description: "状态，0草稿 1已发布"
          format: "int32"
        sub_category_id:
          minimum: 1
          type: "integer"
          description: "子分类ID"
          format: "int32"
      description: "创建文章请求"
    ResultListIdCountVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/IdCountVO"
      description: "统一响应"
    ArticleWithCategoryVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        title:
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章内容"
        user_id:
          type: "integer"
          description: "作者用户ID"
          format: "int64"
        username:
          type: "string"
          description: "作者名称"
        tags:
          type: "string"
          description: "标签，多个用英文逗号分隔"
        status:
          type: "integer"
          description: "状态，0草稿 1已发布"
          format: "int32"
        views:
          type: "integer"
          description: "浏览量"
          format: "int32"
        sub_category_id:
          type: "integer"
          description: "子分类ID"
          format: "int32"
        sub_category_name:
          type: "string"
          description: "子分类名称"
        category_id:
          type: "integer"
          description: "父分类ID"
          format: "int64"
        category_name:
          type: "string"
          description: "父分类名称"
        create_at:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_at:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "带分类的文章"
    ResultListArticleWithCategoryVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/ArticleWithCategoryVO"
      description: "统一响应"
...
    CommentsVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "评论ID"
          format: "int64"
        content:
          type: "string"
          description: "评论内容"
        star:
          type: "number"
          description: "星级评分"
          format: "double"
        username:
          type: "string"
          description: "用户名"
        pic:
          type: "string"
          description: "头像地址"
        article_title:
          type: "string"
          description: "文章标题"
        create_time:
          type: "string"
          description: "创建时间"
        update_time:
          type: "string"
          description: "更新时间"
      description: "评论视图"
    PageVOCommentsVO:
      type: "object"
      properties:
        total:
          type: "integer"
          description: "总记录数"
          format: "int64"
        list:
          type: "array"
          description: "当前页数据"
          items:
            $ref: "#/components/schemas/CommentsVO"
      description: "分页结果"
    ResultPageVOCommentsVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/PageVOCommentsVO"
      description: "统一响应"
...
    ResultArticleWithCategoryVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/ArticleWithCategoryVO"
      description: "统一响应"
    PageVOArticleWithCategoryVO:
      type: "object"
      properties:
        total:
          type: "integer"
          description: "总记录数"
          format: "int64"
        list:
          type: "array"
          description: "当前页数据"
          items:
            $ref: "#/components/schemas/ArticleWithCategoryVO"
      description: "分页结果"
    ResultPageVOArticleWithCategoryVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/PageVOArticleWithCategoryVO"
      description: "统一响应"
...
    ArticleSyncVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        title:
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章正文"
        user_id:
          type: "integer"
          description: "作者用户ID"
          format: "int64"
        tags:
          type: "string"
          description: "标签，多个标签以逗号分隔"
        status:
          type: "integer"
          description: "文章状态，0 草稿，1 已发布"
          format: "int32"
        views:
          type: "integer"
          description: "浏览量"
          format: "int32"
        sub_category_id:
          type: "integer"
          description: "子分类ID"
          format: "int32"
        create_at:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_at:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "文章同步数据"
    ResultListArticleSyncVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/ArticleSyncVO"
      description: "统一响应"
...
    Article:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        title:
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章内容"
        user_id:
          type: "integer"
          description: "作者用户ID"
          format: "int64"
        tags:
          type: "string"
          description: "标签，多个用英文逗号分隔"
        status:
          type: "integer"
          description: "状态，0草稿 1已发布"
          format: "int32"
        views:
          type: "integer"
          description: "浏览量"
          format: "int32"
        sub_category_id:
          type: "integer"
          description: "子分类ID"
          format: "int32"
        create_at:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_at:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "文章"
    PageVOArticle:
      type: "object"
      properties:
        total:
          type: "integer"
          description: "总记录数"
          format: "int64"
        list:
          type: "array"
          description: "当前页数据"
          items:
            $ref: "#/components/schemas/Article"
      description: "分页结果"
    ResultPageVOArticle:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/PageVOArticle"
      description: "统一响应"